refactor: rename developer section to chatbot team section
